Add files via upload
Initial model for Tesla vehicle and Energy Storage and Generation growth. It seems like I need to revise the growth rates for vehicles and Energy Storage and Generation separately. The line for battery growth may not be needed. 

One key point of the initial model is Wright's Law is leads to revenue declines every single time cumulative production doubles by -20%. 

I'm still iffy on how the free cash flow should change from year to year. I have set it to change by -9.1% / +10% a year. Might be too aggressive and may need to switch to a multiplicative model.

The output of the model, the FCF for the first 10 years, and FCF in perpetuity seem reasonable. I think tightening up the growth rates of the vehicle growth is needed.