Add files via upload
I've made solid progress on this model. The parameters for the model have been moved to a separate tab. I have added random variables for Wright's Law and the level of interest rates. I added in the Tesla info and added a placeholder for Nio. I thought it would be cool if we could add in all the EV manufacturers together and see how the industry might grow. Then, you can compare it with individual growth patterns. Next step will be to save the model output each time it is run, and use it to approximate a statistical model. I'm thinking gamma, I don't know why. I'm thinking how to credibility weight the model too, with my model results and reality.